Move dashboards to dashboard app
