+++ amy/workshops/urls.py
@@ -4,16 +4,6 @@ from django.views.generic import RedirectView
 from workshops import views
 
 urlpatterns = [
-    url(r'^$', views.dispatch, name='dispatch'),
-
-    # admin dashboard main page
-    url(r'^admin-dashboard/$', views.admin_dashboard, name='admin-dashboard'),
-
-    # trainee dashboard and trainee-available views
-    url(r'^trainee-dashboard/$', views.trainee_dashboard, name='trainee-dashboard'),
-    url(r'^trainee-dashboard/training_progress/$', views.training_progress, name='training-progress'),
-    url(r'^autoupdate_profile/$', views.autoupdate_profile, name='autoupdate_profile'),
-
     # utility views
     url(r'^search/$', views.search, name='search'),
     url(r'^log/$', views.changes_log, name='changes_log'),
